+++ rdr_service/model/database.py
@@ -85,7 +85,7 @@ from rdr_service.model.study_nph import Participant
 from rdr_service.model.study_nph_sms import SmsSample, SmsJobRun
 from rdr_service.model.rex import Study
 from rdr_service.model.ppsc import Participant
-from rdr_service.model.ppsc_data_transfer import PPSCCore
+from rdr_service.model.ppsc_partner_data_transfer import PPSCCore
 from rdr_service.etl.model.etl_vocabulary import Concept, ConceptRelationship, CombinedQuestionFilter,\
     CombinedSurveyFilter, SourceToConceptMap
 from rdr_service.model.exposomics import ExposomicsM0
